requirements.txt
numpy
networkx
faiss-cpu
sentence-transformers
transformers
yake